+++ src/Consumers/NullBatchConsumer.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Junges\Kafka\Consumers;
+
+use Illuminate\Support\Collection;
+use Junges\Kafka\Contracts\CanConsumeBatchMessages;
+
+class NullBatchConsumer implements CanConsumeBatchMessages
+{
+    /**
+     * {@inheritdoc}
+     */
+    public function handle(Collection $collection): void
+    {
+    }
+}