+++ src/Contracts/CanProduceMessages.php
@@ -4,6 +4,7 @@ namespace Junges\Kafka\Contracts;
 
 use Junges\Kafka\Config\Sasl;
 use Junges\Kafka\Message\Message;
+use Junges\Kafka\Producers\MessageBatch;
 
 interface CanProduceMessages
 {
@@ -110,4 +111,6 @@ interface CanProduceMessages
      * @return bool
      */
     public function send(): bool;
+
+    public function sendBatch(MessageBatch $messageBatch): int;
 }