+++ tests/KafkaFakeTest.php
@@ -6,6 +6,7 @@ use Illuminate\Support\Str;
 use InvalidArgumentException;
 use Junges\Kafka\Facades\Kafka;
 use Junges\Kafka\Message\Message;
+use Junges\Kafka\Producers\MessageBatch;
 use Junges\Kafka\Support\Testing\Fakes\KafkaFake;
 use PHPUnit\Framework\Constraint\ExceptionMessage;
 use PHPUnit\Framework\ExpectationFailedException;
@@ -208,5 +209,21 @@ class KafkaFakeTest extends LaravelKafkaTestCase
         $this->fake->publishOn('undefined')
             ->withBodyKey('foo', 'bar')
             ->send();
+    }
+
+    public function testPublishMessageBatch()
+    {
+        $messageBatch = new MessageBatch();
+        $messageBatch->push(new Message());
+        $messageBatch->push(new Message());
+        $messageBatch->push(new Message());
+
+        $producer = $this->fake->publishOn('default')
+            ->onTopic('topic')
+            ->withBodyKey('test', ['test'])
+            ->withHeaders(['custom' => 'header'])
+            ->withKafkaKey(Str::uuid()->toString());
+
+        $this->assertEquals(3, $producer->sendBatch($messageBatch));
     }
 }